Added Ability to View Products on Specific Aisle
Since it might be hard to figure out which phyisical product
is associated with which entry on the table, we implement a
feature that allows user to see all the products inside of an
aisle. The idea is that if the customer service ever get confused,
they can look at aisle view to figure out which one. Another way
would be to label the physical product with its SKU (variant id)